+++ HISTORY.rst
@@ -2,6 +2,21 @@ History
 -------
 
 
+0.1.1 (2011-11-24)
+++++++++++++++++++
+
+* use pyflakes to find several bugs
+* changed setup.py
+* published to pypi.python.org
+* changed the api_key and app_secret to sys ENV
+* add 在线订购 API
+* add 主动业务通知 API
+* add 淘客 API
+* add 分销 API
+* add 系统时间 API
+* add 评价 API
+
+
 0.1.0 (2011-11-18)
 ++++++++++++++++++